refactor: 💡 dividend checkpoint checks
