+++ src/contract_wrappers/modules/checkpoint/ether_dividend_checkpoint_wrapper.ts
@@ -257,7 +257,14 @@ export default class EtherDividendCheckpointWrapper extends DividendCheckpointWr
       ...params.txData,
       value: valueToWei(params.value, await this.getDecimals()),
     };
-    await this.checkIfDividendIsValid(params.value, params.expiry, params.maturity, params.name);
+    await this.checkIfDividendCreationIsValid(
+      params.expiry,
+      params.maturity,
+      params.value,
+      params.name,
+      undefined,
+      txPayableData,
+    );
     return (await this.contract).createDividend.sendTransactionAsync(
       dateToBigNumber(params.maturity),
       dateToBigNumber(params.expiry),
@@ -273,7 +280,15 @@ export default class EtherDividendCheckpointWrapper extends DividendCheckpointWr
       ...params.txData,
       value: valueToWei(params.value, await this.getDecimals()),
     };
-    await this.checkIfDividendIsValid(params.value, params.expiry, params.maturity, params.name, params.checkpointId);
+    await this.checkIfDividendCreationIsValid(
+      params.expiry,
+      params.maturity,
+      params.value,
+      params.name,
+      undefined,
+      txPayableData,
+      params.checkpointId,
+    );
     return (await this.contract).createDividendWithCheckpoint.sendTransactionAsync(
       dateToBigNumber(params.maturity),
       dateToBigNumber(params.expiry),
@@ -290,11 +305,13 @@ export default class EtherDividendCheckpointWrapper extends DividendCheckpointWr
       ...params.txData,
       value: valueToWei(params.value, await this.getDecimals()),
     };
-    await this.checkIfDividendIsValid(
-      params.value,
+    await this.checkIfDividendCreationIsValid(
       params.expiry,
       params.maturity,
+      params.value,
       params.name,
+      undefined,
+      txPayableData,
       undefined,
       params.excluded,
     );
@@ -316,11 +333,13 @@ export default class EtherDividendCheckpointWrapper extends DividendCheckpointWr
       ...params.txData,
       value: valueToWei(params.value, await this.getDecimals()),
     };
-    await this.checkIfDividendIsValid(
-      params.value,
+    await this.checkIfDividendCreationIsValid(
       params.expiry,
       params.maturity,
+      params.value,
       params.name,
+      undefined,
+      txPayableData,
       params.checkpointId,
       params.excluded,
     );
@@ -380,65 +399,5 @@ export default class EtherDividendCheckpointWrapper extends DividendCheckpointWr
       EtherDividendCheckpoint.abi,
     );
     return logs;
-  };
-
-  private checkIfDividendIsValid = async (
-    value: BigNumber,
-    expiry: Date,
-    maturity: Date,
-    name: string,
-    checkpointId?: number,
-    excluded?: string[],
-  ) => {
-    if (excluded !== undefined) {
-      excluded.forEach(address => assert.isNonZeroETHAddressHex('excluded', address));
-      assert.areThereDuplicatedStrings('excluded', excluded);
-      assert.assert(excluded.length <= EXCLUDED_ADDRESS_LIMIT, 'Too many addresses excluded');
-    }
-    assert.assert(expiry > maturity, 'Expiry before maturity');
-    assert.isFutureDate(expiry, 'Expiry in past');
-    assert.isBigNumberGreaterThanZero(value, 'No dividend sent');
-    const stContract = await this.securityTokenContract();
-    if (checkpointId !== undefined) {
-      const currentCheckpointId = await stContract.currentCheckpointId.callAsync();
-      assert.assert(checkpointId < new BigNumber(currentCheckpointId).toNumber(), 'Invalid checkpoint');
-    }
-    assert.assert(name.length > 0, 'The name can not be empty');
-
-    function checkExcludedAssertions(addr: string) {
-      assert.isNonZeroETHAddressHex('Excluded Address', addr);
-    }
-    if (excluded) {
-      excluded.map(checkExcludedAssertions);
-      assert.areThereDuplicatedStrings('Excluded Addresses', excluded);
-    }
-    let currentSupply;
-    let excludedSupply = new BigNumber(0);
-    if (checkpointId) {
-      currentSupply = await stContract.totalSupplyAt.callAsync(new BigNumber(checkpointId));
-      if (excluded) {
-        const promises = [];
-        for (let i = 0; i < excluded.length; i += 1) {
-          excludedSupply = excludedSupply.plus(
-            promises.push(stContract.balanceOfAt.callAsync(excluded[i], new BigNumber(checkpointId))),
-          );
-        }
-        excludedSupply = BigNumber.sum.apply(null, await Promise.all(promises));
-      }
-    } else {
-      currentSupply = await stContract.totalSupply.callAsync();
-      if (excluded) {
-        const promises = [];
-        for (let i = 0; i < excluded.length; i += 1) {
-          excludedSupply = excludedSupply.plus(promises.push(stContract.balanceOf.callAsync(excluded[i])));
-        }
-        excludedSupply = BigNumber.sum.apply(null, await Promise.all(promises));
-      }
-    }
-    assert.assert(!currentSupply.isZero(), 'Invalid supply, must be greater than 0');
-    assert.assert(
-      currentSupply.isGreaterThan(excludedSupply),
-      'Invalid supply, current supply must be greater than excluded supply',
-    );
   };
 }
